Initial commit, imported from internal Dreamhost repos.
This is the first public commit for the check_s3_health tool.

I wrote it while working for Dreamhost, to run configurable checks on
the health of Ceph RGW instances (Dreamhost DreamObjects).

Dreamhost has given permission for this code to be open-sourced, but
also retains rights over the code (they have an unrestricted license to
do whatever they want with it).

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>
Signed-off-by: Robin H. Johnson <robin.johnson@dreamhost.com>
Signed-off-by: Jeremy Hanmer <jeremy@dreamhost.com>
Signed-off-by: Patrick Lane <patrick@dreamhost.com>
Signed-off-by: Ian Clark <ian@f85.net>